Refactor; introduce channels
